chore!: drop node@12 support
BREAKING CHANGE: node@12 not supported
Use node@14.19 or newer.
