+++ examples/README.md
@@ -13,7 +13,7 @@ Create a [new issue](https://github.com/aeternity/aepp-sdk-js/issues/new) to sug
 - [Wallet WebExtension](browser/wallet-web-extension)
 
 ## NodeJS
-1. [Contract interaction](node/contract-interaction.js)
-2. [Transfer AE](node/transfer-ae.js)
-3. [Paying for spend tx](node/paying-for-spend-tx.js)
-4. [Paying for contract call tx](node/paying-for-contract-call-tx.js)
+1. [Contract interaction](node/contract-interaction.mjs)
+2. [Transfer AE](node/transfer-ae.mjs)
+3. [Paying for spend tx](node/paying-for-spend-tx.mjs)
+4. [Paying for contract call tx](node/paying-for-contract-call-tx.mjs)
