+++ package.json
@@ -14,7 +14,7 @@
     "build:api": "npx autorest tooling/autorest/compiler.yaml --output-artifact:code-model-v4 && npx autorest tooling/autorest/node.yaml --output-artifact:code-model-v4",
     "build": "npm run build:api && webpack --mode=production && npm run build:es",
     "build:dev": "webpack --mode=development && npm run build:es",
-    "docs:examples": "node tooling/docs/examples-to-md.js examples/node/*.js",
+    "docs:examples": "node tooling/docs/examples-to-md.js examples/node/*.mjs",
     "docs:api": "typedoc",
     "commitlint": "commitlint --from develop",
     "lint": "eslint . --ext .ts,.js,.mjs",
@@ -114,10 +114,25 @@
     "src",
     "es"
   ],
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead",
-    "not ie 11"
-  ]
+  "browserslist": {
+    "node": [
+      "node >= 14.19"
+    ],
+    "browser": [
+      "> 1%",
+      "last 2 versions",
+      "not dead",
+      "not ie 11"
+    ],
+    "production": [
+      "node >= 14.19",
+      "> 1%",
+      "last 2 versions",
+      "not dead",
+      "not ie 11"
+    ]
+  },
+  "engines": {
+    "node": ">=14.19.0"
+  }
 }
