+++ webpack.config.js
@@ -2,8 +2,10 @@ const webpack = require('webpack');
 const path = require('path');
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
 const { dependencies } = require('./package.json');
+const babelConfig = require('./babel.config');
 
 function configure(filename, opts = {}) {
+  const isNode = opts.target.includes('node');
   return (env, argv) => ({
     entry: './src/index.ts',
     mode: 'development', // automatically overriden by production flag
@@ -14,11 +16,12 @@ function configure(filename, opts = {}) {
           test: /\.(js|ts)$/,
           include: path.resolve(__dirname, 'src'),
           loader: 'babel-loader',
+          options: { ...babelConfig, browserslistEnv: opts.target.split(':')[1] },
         },
       ],
     },
     optimization: {
-      minimize: opts.target !== 'node',
+      minimize: !isNode,
     },
     resolve: {
       extensions: ['.ts', '.js'],
@@ -27,7 +30,7 @@ function configure(filename, opts = {}) {
       },
     },
     plugins: [
-      ...opts.target === 'node'
+      ...isNode
         ? []
         : [new webpack.ProvidePlugin({
           process: 'process',
@@ -58,7 +61,7 @@ function configure(filename, opts = {}) {
 }
 
 module.exports = [
-  configure('aepp-sdk.js', { target: 'node' }),
-  configure('aepp-sdk.browser.js'),
-  configure('aepp-sdk.browser-script.js', { externals: undefined }),
+  configure('aepp-sdk.js', { target: 'browserslist:node' }),
+  configure('aepp-sdk.browser.js', { target: 'browserslist:browser' }),
+  configure('aepp-sdk.browser-script.js', { target: 'browserslist:browser', externals: undefined }),
 ];
